feature: added tests

+++ src/storage_engine/pager/db_header.rs
@@ -7,7 +7,7 @@ pub struct DatabaseHeader {
     /// Magic string to identify the database file format.
     pub magic_header: [u8; MAGIC_HEADER_STRING.len()],
     /// Size of each page in the database.
-    pub page_size: u32,
+    pub page_size: usize,
 }
 
 impl Default for DatabaseHeader {
